+++ xds/internal/resolver/watch_service.go
@@ -36,7 +36,7 @@ import (
 type serviceUpdate struct {
 	// virtualHost contains routes and other configuration to route RPCs.
 	virtualHost *xdsresource.VirtualHost
-	// clusterSpecifierPlugins contain the configurations for any cluster
+	// clusterSpecifierPlugins contains the configurations for any cluster
 	// specifier plugins emitted by the xdsclient.
 	clusterSpecifierPlugins map[string]clusterspecifier.BalancerConfig
 	// ldsConfig contains configuration that applies to all routes.
